Revert-ish exec commands in shell
This reverts the change which invokes everything in shell and
`shlex.quote` removal, which we need back since we don't want the shell
language to be expanded. Should work:

    ls /etc/passwd
    bash -c "ls /etc/*sswd"

Should fail:

    ls /etc/*

We also require commands to be passed as list, which we do in packit
(function `get_commands_for_actions`) not as a string.

We as well keeps the additional tests and the `set -e` change.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>