exec output: read while the conn is open
we were naive...

...actually, I was

    ws_client.run_forever(timeout=WEBSOCKET_CALL_TIMEOUT)
    errors = ws_client.read_channel(ERROR_CHANNEL)
    response += ws_client.read_all()

My original thinking was that once these 3 calls finish, our exec is
done and we get its output.

WRONG

This only waits for 30 seconds and then halts - so when a command takes
more than 30 seconds (hello npm) to finish, the stream of data is just
cut, no exception raised, nothing suspicious in the logs, and Tomas is
freaking out how come the tarball is not in the repo.

this commit implements:
* read the data while the connection is open
* tests: actually make sure that when we create a srpm from the
  cockpit-podman checkout the file is in there

Resolves https://github.com/packit/packit-service/issues/917

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>