exec: enable setting cwd
honestly, this feature is close to useless since one can easily achieve
this by `bash -c "cd ./subdir && ..."` or `bash -c "/etc..."`

I only implemented it because it was trivial and this is how our
CommandHandler in packit is defined so, potentially, this could be used.

Having this implemented won't give us surprises in future and clearly
defines the interface for other backends (e.g. plain containers)

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>