exec: command fails: copy data out gracefully
When the pod dies and exec is still happening, docker/openshift kill the
exec with rc=137. Copying data after this happens ends with:

https://prod.packit.dev/srpm-build/16278/logs

    WARNING: cannot use rsync: rsync not available in container

Would be nice if oc actually told us the pod is dead and not that
useless warning.

So instead of breaking there, we eat the exception and continue
execution to raise SandcastleCommandFailed which contains all the data
regarding the failure.

Those "weird" changes in api.py are done for sake of the new test case.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>