Don't set locale
To avoid: "ERROR: Ansible requires the locale encoding to be UTF-8; Detected None."
after bumping base image to F37.

It was added in 5e37bb93d to silence autotools rants.
At that time we were setting also LANG, which we don't anymore,
so let's stop setting LC_ALL as well and see if we still need
any locale tweaking.