fix: do not log failed command
When command fails, do not log it, since it is raised with the
exception to the caller, and spam the logs.

!!! Caller should handle this situation on their end, do not depend on
the logs outputted in the API !!!

Removing it, since the exception is caught on the Packit' side and logs
are outputted, therefore this logging is redundant and would cause the
logs to be printed out »twice« in the worker logs.

Related to packit/packit-service#1971
Merge before packit/packit#1939

Signed-off-by: Matej Focko <mfocko@redhat.com>