Add nodejs-packaging-bundler to sandbox
Fedora's Node.js packaging guidelines require that node module
dependencies of a package be bundled up in tarballs that are
provided as separate Sources from the main package's source. It also
recommends the nodejs-packaging-bundler package which provides scripts
to automate this.

To automate this workflow with Packit, it is necessary to run
nodejs-packaging-bundler as part of post-upstream-clone, which runs in
the Sandcastle environment.

Therefore, add nodejs-packaging-bundler to the Sandscastle environment,
for all packages, since there is no way yet to customize this from a
.packit.yaml . The "jowl" package motivated this change, but it should
be useful for any package with node module dependencies.