fix: do not switch to workdir before running ansible
When running Ansible, Python can create its own cache which can clash
with the cache created during user's action in the sandbox. Namely
creating such directory can easily result in broken permissions, since
during the time of the creation of the image, it is being done with the
highest permissions of `root`.

Therefore move the permission-fix and also switch of the work directory
**after** the Ansible is run.

Related to cockpit-project/cockpit-machines#1350

Signed-off-by: Matej Focko <mfocko@redhat.com>